edit option in the given list

+++ app.js
@@ -1,6 +1,6 @@
 // functionality that we are going to perform 
 
-// Book Class: Represents a Book
+// Book Class: Represents a Book.........................................................
 class Book {
     constructor(title, author, isbn) {
       this.title = title;
@@ -9,7 +9,7 @@ class Book {
     }
   }
   
-  // UI Class: Handle UI Tasks
+  // UI Class: Handle UI Tasks...............................................................
   class UI {
     static displayBooks() {
         // const storedBooks =[
@@ -38,7 +38,8 @@ class Book {
         <td>${book.title}</td>
         <td>${book.author}</td>
         <td>${book.isbn}</td>
-        <td><a href="#" class="btn btn-danger btn-sm delete">X</a></td>
+        <td><a href="#" class="btn btn-danger btn-sm delete" id="delete">X</a></td>
+        <td><button class="btn btn-success btn-sm" id="edit" onclick="editBook(${book.isbn})">EDIT</button></td>
       `;
   
       list.appendChild(row);
@@ -69,7 +70,23 @@ class Book {
     }
   }
   
-  // Store Class: Handles Storage
+  function editBook(isbn) {
+    // Set Button Text
+    document.getElementById('addBook').value="Edit Book";
+    // Get Books
+    const books = Store.getBooks();
+    // Assign values to there respective fields
+    books.forEach((book, index) => {
+      // console.log(book.title +" "+ typeof book.isbn+" "+typeof isbn);
+      if(book.isbn == isbn) {
+        // console.log(book.title);
+        document.getElementById('title').value=book.title;
+        document.getElementById('author').value=book.author;
+        document.getElementById('isbn').value=book.isbn;
+      }
+    });
+  }
+  // Store Class: Handles Storage....................................................
   class Store {
     static getBooks() {
       let books;
@@ -83,9 +100,31 @@ class Book {
     }
   
     static addBook(book) {
+      // console.log(book);
       const books = Store.getBooks();
       books.push(book);
       localStorage.setItem('books', JSON.stringify(books));
+    }
+
+    static editBook(editData) {
+      const books = Store.getBooks();
+      let edited = false;
+      books.forEach((book, index) => {
+        // console.log(book.title +" "+ typeof book.isbn+" "+typeof isbn);
+        if(book.isbn == editData.isbn) {
+          edited = true;
+          books.splice(index, 1);
+          books.splice(index, 0, editData);
+        }
+      });
+      if (!edited){
+        books.push(editData);
+        localStorage.setItem('books', JSON.stringify(books));
+        UI.showAlert('ISBN update Added New Book', 'success');
+      } else {
+        localStorage.setItem('books', JSON.stringify(books));
+        UI.showAlert('Details Updated', 'success');
+      }
     }
   
     static removeBook(isbn) {
@@ -100,49 +139,75 @@ class Book {
       localStorage.setItem('books', JSON.stringify(books));
     }
   }
+  // console.log(localStorage);
   
-  // Event: Display Books
+  // Event: Display Books.........................................................................
   document.addEventListener('DOMContentLoaded', UI.displayBooks);
   
-  // Event: Add a Book
+  // Event: Add a Book............................................................................
   document.querySelector('#book-form').addEventListener('submit', (e) => {
     // Prevent actual submit
     e.preventDefault();
-  
-    // Get form values
-    const title = document.querySelector('#title').value;
-    const author = document.querySelector('#author').value;
-    const isbn = document.querySelector('#isbn').value;
-  
-    // Validate
-    if(title === '' || author === '' || isbn === '') {
-      UI.showAlert('Please fill in all fields', 'danger');
-    } else {
-      // Instatiate book
-      const book = new Book(title, author, isbn);
-  
-      // Add Book to UI
-      UI.addBookToList(book);
-  
-      // Add book to store
-      Store.addBook(book);
-  
-      // Show success message
-      UI.showAlert('Book Added Successfully', 'success');
-  
-      // Clear fields
-      UI.clearFields();
+
+    if (document.getElementById('addBook').value === 'Add Book'){
+      // Get form values
+      const title = document.querySelector('#title').value;
+      const author = document.querySelector('#author').value;
+      const isbn = document.querySelector('#isbn').value;
+    
+      // Validate
+      if(title === '' || author === '' || isbn === '') {
+        UI.showAlert('Please fill in all fields', 'danger');
+      } else {
+        // Instatiate book
+        const book = new Book(title, author, isbn);
+    
+        // Add Book to UI
+        UI.addBookToList(book);
+    
+        // Add book to store
+        Store.addBook(book);
+    
+        // Show success message
+        UI.showAlert('Book Added Successfully', 'success');
+    
+        // Clear fields
+        UI.clearFields();
+      }
+    } else if (document.getElementById('addBook').value === 'Edit Book'){
+      const title = document.querySelector('#title').value;
+      const author = document.querySelector('#author').value;
+      const isbn = document.querySelector('#isbn').value;
+    
+      // Validate
+      if(title === '' || author === '' || isbn === '') {
+        UI.showAlert('Please fill in all fields', 'danger');
+      }else {
+        const book = new Book(title,author,isbn);
+
+        //Edit Book in UI
+
+        //Edit Book Store
+        Store.editBook(book);
+
+        UI.displayBooks();
+      }
+      // console.log("it's Edit");
     }
   });
   
-  // Event: Remove a Book
+  // Event: Remove a Book................................................................................
   document.querySelector('#book-list').addEventListener('click', (e) => {
-    // Remove book from UI
-    UI.deleteBook(e.target);
-  
-    // Remove book from store
-    Store.removeBook(e.target.parentElement.previousElementSibling.textContent);
-  
-    // Show success message
-    UI.showAlert('Book Removed', 'success');
+    const action = e.target.id;
+
+    if (action === 'delete') {
+      // Remove book from UI
+      UI.deleteBook(e.target);
+    
+      // Remove book from store
+      Store.removeBook(e.target.parentElement.previousElementSibling.textContent);
+    
+      // Show success message
+      UI.showAlert('Book Removed', 'success');
+    }
   });
